Created the tests/utilities.py file to hold common features.
